Add optional arg to login/configure script for env
When using this role for baking the gremlin agent into an image, you
may not know the environment (or another identifying contextual bit),
so support passing an optional arg when calling the login/configure
script (e.g. in AWS user data), that will append the contextual bit to
the templated (and baked) service name.

Tracking: #3